feat: refactor extract controller
