+++ src/admin/admin.controller.ts
@@ -6,7 +6,6 @@ import {
   Get,
   HttpException,
   Param,
-  ParseIntPipe,
   Patch,
   Post,
   Request,
@@ -15,16 +14,11 @@ import {
 } from '@nestjs/common';
 import { UserService } from '../user/user.service';
 import { AdminService } from './admin.service';
-import { JwtAuthGuard } from 'src/auth/guard/jwt-auth.guard';
-import { HypePermission, HypeRole, User } from '../entity';
+import { User } from '../entity';
 import { PermissionGuard } from '../auth/guard/permission.guard';
 import { Permissions } from '../auth/permission.decorator';
-import { InjectModel } from '@nestjs/sequelize';
 import {
-  AdminAssignUserRoleRequest,
   AdminChangePasswordRequest,
-  AdminCreatePermissionRequest,
-  AdminCreateRoleRequest,
   AdminCreateUserRequest,
   AdminUpdateUserRequest,
   UserResponseModel,
@@ -37,10 +31,6 @@ export class AdminController {
   constructor(
     private userService: UserService,
     private adminService: AdminService,
-    @InjectModel(HypePermission)
-    private hypePermission: typeof HypePermission,
-    @InjectModel(HypeRole)
-    private hypeRole: typeof HypeRole,
   ) {}
 
   @UseGuards(PermissionGuard)
@@ -124,19 +114,6 @@ export class AdminController {
 
   @UseGuards(PermissionGuard)
   @Permissions('user_management')
-  @Patch('user-roles/:uid')
-  async assignUserRole(
-    @Param('uid', new ParseIntPipe()) uid: number,
-    @Request() req: HypeRequest,
-    @Body() body: AdminAssignUserRoleRequest,
-  ) {
-    if (body.roles != null) {
-      return this.adminService.applyUserRoles(req.user, uid, body.roles);
-    }
-  }
-
-  @UseGuards(PermissionGuard)
-  @Permissions('user_management')
   @Delete('user/:uid')
   async deleteUser(@Param('uid') uid: string): Promise<UserResponseModel> {
     return this.userService.updateUser({
@@ -150,101 +127,5 @@ export class AdminController {
   @Post('init-project')
   async initProject(@Request() req: HypeRequest) {
     return this.adminService.initProject(req.user);
-  }
-
-  @UseGuards(PermissionGuard)
-  @Permissions('permission_management')
-  @Get('roles')
-  async getRoles(): Promise<{ data: Array<HypeRole>; total: number }> {
-    const [data, total] = await Promise.all([
-      this.hypeRole.findAll({
-        where: {
-          deletedAt: null,
-        },
-        include: [],
-      }),
-      this.hypeRole.count({ where: { deletedAt: null } }),
-    ]);
-    return {
-      data,
-      total,
-    };
-  }
-
-  @UseGuards(PermissionGuard)
-  @Permissions('permission_management')
-  @Get('role/:id')
-  async getRole(@Param('id') id: string) {
-    return this.adminService.getRole(parseInt(id));
-  }
-
-  @UseGuards(PermissionGuard)
-  @Permissions('permission_management')
-  @Post('role')
-  async createRole(@Body() body: AdminCreateRoleRequest) {
-    return await this.adminService.creatRole({
-      slug: body.slug,
-      name: body.name,
-      // createdBy: req.user.id,
-      roleType: 'normal',
-    });
-  }
-
-  @UseGuards(PermissionGuard)
-  @Permissions('permission_management')
-  @Delete('role/:id')
-  async deleteRole(@Param('id') id: string, @Request() req: HypeRequest) {
-    return this.adminService.deleteRole(req.user, {
-      id: parseInt(id),
-    });
-  }
-
-  @Get('assign-role/:uid')
-  @UseGuards(PermissionGuard)
-  @Permissions('user_management')
-  async getAssignRole(@Param('uid') id: string) {
-    return this.adminService.getAssignRole(parseInt(id));
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Post('permission/datalist')
-  async getPermissionList(): Promise<{
-    data: Array<HypePermission>;
-    total: number;
-  }> {
-    const [data, total] = await Promise.all([
-      this.hypePermission.findAll({
-        where: {
-          deletedAt: null,
-        },
-        include: [HypeRole],
-      }),
-      this.hypePermission.count({
-        where: { deletedAt: null },
-      }),
-    ]);
-    return {
-      data,
-      total,
-    };
-  }
-
-  @UseGuards(PermissionGuard)
-  @Permissions('permission_management')
-  @Post('permission')
-  async createPermission(@Body() body: AdminCreatePermissionRequest) {
-    return this.adminService.createPermission({
-      name: body.name,
-      slug: body.slug,
-    });
-  }
-
-  @UseGuards(PermissionGuard)
-  @Permissions('permission_management')
-  @Delete('permission/:id')
-  async deletePermission(@Param('id') id: string, @Request() req: HypeRequest) {
-    return this.adminService.deletePermission(req.user, {
-      id: parseInt(id),
-    });
   }
 }
